Move code specific for 2020.3 into common source set
