--- clion/src/203/main/kotlin/org/rust/clion/cargo/CidrBuildConfigurationProvider.kt
@@ -1,8 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.clion.cargo
-
-typealias CidrBuildConfigurationProvider = com.jetbrains.cidr.execution.build.CidrBuildConfigurationProvider
