--- debugger/src/203/main/resources/META-INF/platform-core-debugger-only.xml
@@ -1,2 +1,0 @@
-<idea-plugin>
-</idea-plugin>
