--- debugger/src/203/main/resources/META-INF/platform-native-debug-only.xml
@@ -1,2 +1,0 @@
-<idea-plugin>
-</idea-plugin>
