+++ debugger/src/main/resources/META-INF/native-debug-only.xml
@@ -1,4 +1,3 @@
 <idea-plugin xmlns:xi="http://www.w3.org/2001/XInclude">
     <xi:include href="/META-INF/debugger-only.xml" xpointer="xpointer(/idea-plugin/*)"/>
-    <xi:include href="/META-INF/platform-native-debug-only.xml" xpointer="xpointer(/idea-plugin/*)"/>
 </idea-plugin>
