+++ ml-completion/src/main/resources/META-INF/ml-completion-only.xml
