+++ ml-completion/src/test/kotlin/org/rust/ml/RsModelMetadataConsistencyTest.kt
