+++ ml-completion/src/test/resources/META-INF/plugin.xml
