--- src/203/main/kotlin/org/rust/lang/core/resolve/ref/compatUtils.kt
@@ -1,11 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.lang.core.resolve.ref
-
-import com.intellij.util.containers.HashingStrategy
-
-@Suppress("UnstableApiUsage")
-fun <T> canonicalStrategy(): HashingStrategy<T> = HashingStrategy.canonical()
