+++ src/test/kotlin/org/rust/cargo/RsWithToolchainTestBase.kt
@@ -8,6 +8,9 @@ package org.rust.cargo
 import com.intellij.openapi.util.RecursionManager
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.newvfs.impl.VfsRootAccess
+import com.intellij.testFramework.builders.ModuleFixtureBuilder
+import com.intellij.testFramework.fixtures.CodeInsightFixtureTestCase
+import com.intellij.util.ThrowableRunnable
 import com.intellij.util.ui.UIUtil
 import org.rust.*
 import org.rust.cargo.project.model.impl.testCargoProjects
@@ -20,7 +23,7 @@ import org.rust.openapiext.pathAsPath
  * Unlike [org.rust.RsTestBase] it does not use in-memory temporary VFS
  * and instead copies real files.
  */
-abstract class RsWithToolchainTestBase : RsWithToolchainPlatformTestBase() {
+abstract class RsWithToolchainTestBase : CodeInsightFixtureTestCase<ModuleFixtureBuilder<*>>() {
 
     protected lateinit var rustupFixture: RustupTestFixture
 
@@ -44,7 +47,7 @@ abstract class RsWithToolchainTestBase : RsWithToolchainPlatformTestBase() {
         project.testCargoProjects.discoverAndRefreshSync()
     }
 
-    override fun runTestInternal(context: TestContext) {
+    override fun runTestRunnable(testRunnable: ThrowableRunnable<Throwable>) {
         val skipReason = rustupFixture.skipTestReason ?: getIgnoredInNewResolveReason(project)
         if (skipReason != null) {
             System.err.println("SKIP \"$name\": $skipReason")
@@ -64,7 +67,7 @@ abstract class RsWithToolchainTestBase : RsWithToolchainPlatformTestBase() {
                 return
             }
         }
-        super.runTestInternal(context)
+        super.runTestRunnable(testRunnable)
     }
 
     override fun setUp() {
