--- toml/src/203/test/kotlin/org/rust/toml/completion/CargoTomlCompletionEditionTestBase.kt
@@ -1,9 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.toml.completion
-
-// BACKCOMPAT: 2020.2. Merge it with `CargoTomlCompletionEditionTest`
-abstract class CargoTomlCompletionEditionTestBase : CargoTomlCompletionTestBase()
